Split TWS Python test into multiple small tests to speed up the CI
